requirements.txt
dash==3.2.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-daq==0.6.0
dash-bootstrap-components==2.0.3
httpx
python-dotenv
